pyproject.toml
[tool.poetry]
name = "bia6304-assignment5"
version = "0.1.0"
description = "Assignment 5: Retrieval Augmented Generation"
authors = ["Martin Bukowski <martin.bukowski@rockhurst.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "*"
pandas = "*"
scikit-learn = "*"
numpy = "<2"
groq = "*"
google-generativeai = "*"
python-dotenv = "*"
click = "*"
wonderwords = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
